+++ include/input_stream.h
@@ -27,9 +27,9 @@ class InputStream
 {
 public:
    virtual ~InputStream() {}
-   virtual long read(void *buffer, long length) = 0;
-   virtual bool seek(long offset, int whence) = 0;
-   virtual long tell() = 0;
+   virtual int32_t read(void *buffer, int32_t length) = 0;
+   virtual bool seek(int32_t offset, int whence) = 0;
+   virtual int32_t tell() = 0;
    virtual void close() = 0;
 };
