+++ include/mem_input_stream.h
@@ -34,9 +34,9 @@ public:
    MemInputStream();
    ~MemInputStream();
    void open(void *data, size_t length, bool own_data = true);
-   long read(void *buffer, long length);
-   bool seek(long offset, int whence);
-   long tell();
+   int32_t read(void *buffer, int32_t length);
+   bool seek(int32_t offset, int whence);
+   int32_t tell();
    void close();
 };
